+++ src/Symfony/Component/Translation/Tests/Extractor/PhpExtractorTest.php
@@ -130,6 +130,120 @@ EOF;
         $this->assertEquals(['sources' => [$filename.':37']], $catalogue->getMetadata('other-domain-test-no-params-short-array', 'not_messages'));
     }
 
+    /**
+     * @dataProvider resourcesProvider
+     *
+     * @param array|string $resource
+     */
+    public function testBlankExtraction($resource)
+    {
+        // Arrange
+        $extractor = new PhpExtractor();
+        $extractor->setPrefix('prefix');
+        $extractor->blank(true);
+        $catalogue = new MessageCatalogue('en');
+
+        // Act
+        $extractor->extract($resource, $catalogue);
+
+        $expectedHeredoc = <<<EOF
+heredoc key with whitespace and escaped \$\n sequences
+EOF;
+        $expectedNowdoc = <<<'EOF'
+nowdoc key with whitespace and nonescaped \$\n sequences
+EOF;
+        // Assert
+        $expectedCatalogue = [
+            'messages' => [
+                'translatable single-quoted key' => '',
+                'translatable double-quoted key' => '',
+                'translatable heredoc key' => '',
+                'translatable nowdoc key' => '',
+                "translatable double-quoted key with whitespace and escaped \$\n\" sequences" => '',
+                'translatable single-quoted key with whitespace and nonescaped \$\n\' sequences' => '',
+                'translatable single-quoted key with "quote mark at the end"' => '',
+                'translatable '.$expectedHeredoc => '',
+                'translatable '.$expectedNowdoc => '',
+                'translatable concatenated message with heredoc and nowdoc' => '',
+                'translatable default domain' => '',
+                'translatable-fqn single-quoted key' => '',
+                'translatable-fqn double-quoted key' => '',
+                'translatable-fqn heredoc key' => '',
+                'translatable-fqn nowdoc key' => '',
+                "translatable-fqn double-quoted key with whitespace and escaped \$\n\" sequences" => '',
+                'translatable-fqn single-quoted key with whitespace and nonescaped \$\n\' sequences' => '',
+                'translatable-fqn single-quoted key with "quote mark at the end"' => '',
+                'translatable-fqn '.$expectedHeredoc => '',
+                'translatable-fqn '.$expectedNowdoc => '',
+                'translatable-fqn concatenated message with heredoc and nowdoc' => '',
+                'translatable-fqn default domain' => '',
+                'translatable-short single-quoted key' => '',
+                'translatable-short double-quoted key' => '',
+                'translatable-short heredoc key' => '',
+                'translatable-short nowdoc key' => '',
+                "translatable-short double-quoted key with whitespace and escaped \$\n\" sequences" => '',
+                'translatable-short single-quoted key with whitespace and nonescaped \$\n\' sequences' => '',
+                'translatable-short single-quoted key with "quote mark at the end"' => '',
+                'translatable-short '.$expectedHeredoc => '',
+                'translatable-short '.$expectedNowdoc => '',
+                'translatable-short concatenated message with heredoc and nowdoc' => '',
+                'translatable-short default domain' => '',
+                'single-quoted key' => '',
+                'double-quoted key' => '',
+                'heredoc key' => '',
+                'nowdoc key' => '',
+                "double-quoted key with whitespace and escaped \$\n\" sequences" => '',
+                'single-quoted key with whitespace and nonescaped \$\n\' sequences' => '',
+                'single-quoted key with "quote mark at the end"' => '',
+                $expectedHeredoc => '',
+                $expectedNowdoc => '',
+                'concatenated message with heredoc and nowdoc' => '',
+                'default domain' => '',
+            ],
+            'not_messages' => [
+                'translatable other-domain-test-no-params-short-array' => '',
+                'translatable other-domain-test-no-params-long-array' => '',
+                'translatable other-domain-test-params-short-array' => '',
+                'translatable other-domain-test-params-long-array' => '',
+                'translatable typecast' => '',
+                'translatable-fqn other-domain-test-no-params-short-array' => '',
+                'translatable-fqn other-domain-test-no-params-long-array' => '',
+                'translatable-fqn other-domain-test-params-short-array' => '',
+                'translatable-fqn other-domain-test-params-long-array' => '',
+                'translatable-fqn typecast' => '',
+                'translatable-short other-domain-test-no-params-short-array' => '',
+                'translatable-short other-domain-test-no-params-long-array' => '',
+                'translatable-short other-domain-test-params-short-array' => '',
+                'translatable-short other-domain-test-params-long-array' => '',
+                'translatable-short typecast' => '',
+                'other-domain-test-no-params-short-array' => '',
+                'other-domain-test-no-params-long-array' => '',
+                'other-domain-test-params-short-array' => '',
+                'other-domain-test-params-long-array' => '',
+                'typecast' => '',
+            ],
+        ];
+        $actualCatalogue = $catalogue->all();
+
+        $this->assertEquals($expectedCatalogue, $actualCatalogue);
+
+        $filename = str_replace(\DIRECTORY_SEPARATOR, '/', __DIR__).'/../fixtures/extractor/translatable.html.php';
+        $this->assertEquals(['sources' => [$filename.':2']], $catalogue->getMetadata('translatable single-quoted key'));
+        $this->assertEquals(['sources' => [$filename.':37']], $catalogue->getMetadata('translatable other-domain-test-no-params-short-array', 'not_messages'));
+
+        $filename = str_replace(\DIRECTORY_SEPARATOR, '/', __DIR__).'/../fixtures/extractor/translatable-fqn.html.php';
+        $this->assertEquals(['sources' => [$filename.':2']], $catalogue->getMetadata('translatable-fqn single-quoted key'));
+        $this->assertEquals(['sources' => [$filename.':37']], $catalogue->getMetadata('translatable-fqn other-domain-test-no-params-short-array', 'not_messages'));
+
+        $filename = str_replace(\DIRECTORY_SEPARATOR, '/', __DIR__).'/../fixtures/extractor/translatable-short.html.php';
+        $this->assertEquals(['sources' => [$filename.':2']], $catalogue->getMetadata('translatable-short single-quoted key'));
+        $this->assertEquals(['sources' => [$filename.':37']], $catalogue->getMetadata('translatable-short other-domain-test-no-params-short-array', 'not_messages'));
+
+        $filename = str_replace(\DIRECTORY_SEPARATOR, '/', __DIR__).'/../fixtures/extractor/translation.html.php';
+        $this->assertEquals(['sources' => [$filename.':2']], $catalogue->getMetadata('single-quoted key'));
+        $this->assertEquals(['sources' => [$filename.':37']], $catalogue->getMetadata('other-domain-test-no-params-short-array', 'not_messages'));
+    }
+
     public function testExtractionFromIndentedHeredocNowdoc()
     {
         $catalogue = new MessageCatalogue('en');
@@ -142,6 +256,25 @@ EOF;
             'messages' => [
                 "heredoc\nindented\n  further" => "prefixheredoc\nindented\n  further",
                 "nowdoc\nindented\n  further" => "prefixnowdoc\nindented\n  further",
+            ],
+        ];
+
+        $this->assertEquals($expectedCatalogue, $catalogue->all());
+    }
+
+    public function testBlankExtractionFromIndentedHeredocNowdoc()
+    {
+        $catalogue = new MessageCatalogue('en');
+
+        $extractor = new PhpExtractor();
+        $extractor->setPrefix('prefix');
+        $extractor->blank(true);
+        $extractor->extract(__DIR__.'/../fixtures/extractor-7.3/translation.html.php', $catalogue);
+
+        $expectedCatalogue = [
+            'messages' => [
+                "heredoc\nindented\n  further" => '',
+                "nowdoc\nindented\n  further" => '',
             ],
         ];